move useFormHandlers as hooks in individual file
