+++ src/store/hooks.ts
@@ -4,3 +4,6 @@ import type { AppDispatch, RootState } from './store';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+
+// Export the form handlers hook
+export { useFormHandlers } from './hooks/useFormHandlers';
